+++ examples/run/cpu/apps/throughput_st.cpp
@@ -1,6 +1,6 @@
 /** TRACCC library, part of the ACTS project (R&D line)
  *
- * (c) 2021-2022 CERN for the benefit of the ACTS project
+ * (c) 2021-2024 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -8,11 +8,11 @@
 // Local include(s).
 #include "traccc/examples/throughput_st.hpp"
 
-#include "traccc/examples/full_chain_algorithm.hpp"
+#include "traccc/examples/cpu/full_chain_algorithm.hpp"
 
 int main(int argc, char* argv[]) {
 
     // Execute the throughput test.
-    return traccc::throughput_st<traccc::cuda::full_chain_algorithm>(
-        "Single-threaded CUDA GPU throughput tests", argc, argv);
+    return traccc::throughput_st<traccc::full_chain_algorithm>(
+        "Single-threaded host-only throughput tests", argc, argv);
 }
