+++ examples/run/cpu/apps/truth_finding_example.cpp
