+++ examples/run/cuda/apps/truth_finding_example_cuda.cpp
