+++ examples/run/cuda/apps/truth_fitting_example_cuda.cpp
