+++ examples/run/sycl/CMakeLists.txt
@@ -15,13 +15,13 @@ target_link_libraries( traccc_examples_sycl
           traccc::core traccc::device_common traccc::sycl
           traccc_examples_common )
 
-traccc_add_executable( throughput_st_sycl "src/throughput_st.cpp"
+traccc_add_executable( throughput_st_sycl "apps/throughput_st.cpp"
    LINK_LIBRARIES indicators::indicators vecmem::core vecmem::sycl
                   detray::detectors detray::io traccc::io traccc::performance
                   traccc::core traccc::device_common traccc::sycl
                   traccc::options traccc_examples_sycl )
 
-traccc_add_executable( throughput_mt_sycl "src/throughput_mt.cpp"
+traccc_add_executable( throughput_mt_sycl "apps/throughput_mt.cpp"
    LINK_LIBRARIES indicators::indicators TBB::tbb vecmem::core vecmem::sycl
                   detray::detectors detray::io traccc::io traccc::performance
                   traccc::core traccc::device_common traccc::sycl
